Refactored visit and patient common code to mixin.
